+++ src/Query/IsbnQuery.java
@@ -12,7 +12,7 @@ public class IsbnQuery implements BookQuery {
     @Override
     public ArrayList<Book> search(ArrayList<Book> books, String searchParam){
         ArrayList<Book> newBooks = new ArrayList<>();
-        newBooks.add(Books.getBookHash().get(Integer.parseInt(searchParam)));
+        newBooks.add(Books.getBookHash().get(Long.parseLong(searchParam)));
         return newBooks;
     }
 }
